REFACTOR: Split some of the skills constants into multiple files - Will make it easier to create relations
