+++ database/skills/technicalHardSkills/technicalHardSkillsDatabases.ts
@@ -89,81 +89,6 @@ export const redis: Skill = {
   ],
 };
 
-export const prisma: Skill = {
-  name: "Prisma",
-  category: "Object Relational Mappers",
-  isMainSkill: true,
-  skillType: "hard",
-  slug: "prisma-orm",
-  technicalGeneralSkills: [
-    databaseManagementSystems,
-    databases,
-    sql,
-    indexing,
-    orm,
-  ],
-};
-
-export const drizzle: Skill = {
-  name: "Drizzle",
-  category: "Object Relational Mappers",
-  isMainSkill: true,
-  skillType: "hard",
-  slug: "drizzle-orm",
-  technicalGeneralSkills: [
-    databaseManagementSystems,
-    databases,
-    sql,
-    indexing,
-    orm,
-  ],
-};
-
-export const mongoose: Skill = {
-  name: "Mongoose",
-  category: "Object Relational Mappers",
-  isMainSkill: true,
-  skillType: "hard",
-  slug: "mongoose-orm",
-  technicalGeneralSkills: [
-    databaseManagementSystems,
-    databases,
-    noSql,
-    indexing,
-    orm,
-  ],
-};
-
-export const sqlalchemy: Skill = {
-  name: "SQLAlchemy",
-  category: "Object Relational Mappers",
-  isMainSkill: true,
-  skillType: "hard",
-  slug: "sqlalchemy",
-  technicalGeneralSkills: [
-    databaseManagementSystems,
-    databases,
-    sql,
-    indexing,
-    orm,
-  ],
-};
-
-export const hibernate: Skill = {
-  name: "Hibernate",
-  category: "Object Relational Mappers",
-  isMainSkill: true,
-  skillType: "hard",
-  slug: "hibernate-orm",
-  technicalGeneralSkills: [
-    databaseManagementSystems,
-    databases,
-    sql,
-    indexing,
-    orm,
-  ],
-};
-
 const technicalHardSkillsDatabases: Skill[] = [
   normalisation,
   postgreSQL,
@@ -171,11 +96,6 @@ const technicalHardSkillsDatabases: Skill[] = [
   sqlite,
   mongoDB,
   redis,
-  prisma,
-  drizzle,
-  mongoose,
-  sqlalchemy,
-  hibernate,
 ];
 
 export default technicalHardSkillsDatabases;
