+++ database/skills/technicalHardSkills/technicalHardSkillsOthers.ts
@@ -1,13 +1,6 @@
 import Skill from "@/types/skills";
 import { gameDevelopment } from "../generalSkills";
 
-export const zod: Skill = {
-  name: "Zod",
-  category: "Core Computer Science",
-  skillType: "hard",
-  slug: "zod",
-};
-
 export const gameMakerStudio: Skill = {
   name: "GameMaker Studio 2",
   category: "Game Development",
@@ -16,6 +9,6 @@ export const gameMakerStudio: Skill = {
   technicalGeneralSkills: [gameDevelopment],
 };
 
-const technicalHardSkillsGameDev: Skill[] = [zod, gameMakerStudio];
+const technicalHardSkillsGameDev: Skill[] = [gameMakerStudio];
 
 export default technicalHardSkillsGameDev;
